refactor(placement): split virtual-screen migration out of lifecycle.cpp
